Refactor SO into class
